feat(orchestrator): Issue 27 - DelegateController extracted from delegate-tool.ts
